Storage html added, controller added
