+++ WareHouseMgmt/src/main/java/com/dbs/service/StorageService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 /**
  * @author Pankesh
- *
+ * Service for Storage related operations
  */
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
